refactor - 매직넘버 제거
