+++ src/main/java/baseball/controller/Game.java
@@ -2,7 +2,7 @@ package baseball.controller;
 
 import baseball.domain.Balls;
 import baseball.domain.Computer;
-import baseball.domain.RandomNumbers;
+import baseball.util.RandomNumbers;
 import baseball.domain.Referee;
 import baseball.view.Input;
 import baseball.view.Output;
@@ -18,7 +18,7 @@ public class Game {
 
     public void setGame() {
         Output.printStartGameMessage();
-        Balls comBalls = Balls.createBalls(RandomNumbers.createRandomBalls());
+        Balls comBalls = Balls.createBalls(RandomNumbers.createRandomNumbers());
         referee = new Referee(new Computer(comBalls));
     }
